✨ Feat: id 자료형 변경(Long -> Int)
#


Former-commit-id: 3b559c3e31d1cf7e235a6ee132862056e2400f1a

+++ Classik_Backend/src/main/kotlin/com/ssafy/Classik_Backend/repository/TrackRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface TrackRepository : JpaRepository<Track, Long> {
+interface TrackRepository : JpaRepository<Track, Int> {
 
     fun findAllByTitle(title: String): List<Track>
